Extract form object behavior
