+++ sig/structured_params/params.rbs
@@ -49,7 +49,6 @@ module StructuredParams
   #     <%= f.text_field :name %>
   #     <%= f.text_field :email %>
   #   <% end %>
-  # rubocop:disable Metrics/ClassLength
   class Params
     include ActiveModel::Model
 
@@ -61,21 +60,16 @@ module StructuredParams
 
     include I18n
 
+    include FormObject
+
     @errors: ::StructuredParams::Errors?
 
     self.@structured_attributes: Hash[Symbol, singleton(::StructuredParams::Params)]?
 
     self.@model_name: ::ActiveModel::Name?
 
-    # Override model_name for form helpers
-    # By default, removes "Parameters", "Parameter", or "Form" suffix from class name
-    # This allows the class to work seamlessly with Rails form helpers
-    #
-    # Example:
-    #   UserRegistrationForm.model_name.name       # => "UserRegistration"
-    #   UserRegistrationForm.model_name.param_key  # => "user_registration"
-    #   UserParameters.model_name.name             # => "User"
-    #   Admin::UserForm.model_name.name            # => "Admin::User"
+    # Override model_name for Rails integrations.
+    # By default, removes "Parameters", "Parameter", or "Form" suffix from class name.
     # : () -> ::ActiveModel::Name
     def self.model_name: () -> ::ActiveModel::Name
 
@@ -115,16 +109,6 @@ module StructuredParams
     # : () -> ::StructuredParams::Errors
     def errors: () -> ::StructuredParams::Errors
 
-    # Form object support for Rails helpers.
-    # : () -> bool
-    def persisted?: () -> bool
-
-    # : () -> nil
-    def to_key: () -> nil
-
-    # : () -> self
-    def to_model: () -> self
-
     # Convert structured objects to Hash and get attributes
     # : (?symbolize: false, ?compact_mode: :none | :nil_only | :all_blank) -> Hash[String, untyped]
     # : (?symbolize: true, ?compact_mode: :none | :nil_only | :all_blank) -> Hash[Symbol, untyped]
@@ -139,35 +123,6 @@ module StructuredParams
 
     # : (ActionController::Parameters) -> Hash[untyped, untyped]
     def process_action_controller_parameters: (ActionController::Parameters) -> Hash[untyped, untyped]
-
-    # Whether to call params.require(param_key) before permitting.
-    #
-    # Only ever true for Form-suffixed classes (form_class?); non-Form
-    # Params/Parameters subclasses always permit the top level directly.
-    #
-    # - A value nested under the model's param_key (e.g. params[:user_registration])
-    #   always wins, even if params were already permitted higher up, since that
-    #   inner key still needs to be require()'d out.
-    # - Otherwise, params that are already permitted, or whose shape looks flat
-    #   (see flat_parameters?), are used as-is without requiring.
-    # - Anything else falls through to true, so params.require raises
-    #   ActionController::ParameterMissing instead of guessing which keys
-    #   belong to this form.
-    # : (ActionController::Parameters) -> bool
-    def require_nested_parameters?: (ActionController::Parameters) -> bool
-
-    # : (ActionController::Parameters) -> bool
-    def matches_model_name?: (ActionController::Parameters) -> bool
-
-    # Only treat params as already-flat attributes when none of the top-level
-    # values are themselves nested. A nested value under some other key means
-    # the request shape is ambiguous, so we fall through to raising
-    # ParameterMissing instead of silently guessing which keys belong here.
-    # : (ActionController::Parameters) -> bool
-    def flat_parameters?: (ActionController::Parameters) -> bool
-
-    # : (ActionController::Parameters) -> bool
-    def attribute_keys_present?: (ActionController::Parameters) -> bool
 
     # Execute structured parameter validation
     # : () -> void
